feat(settings): move agent context diagnostics from Connect to Debug
The Run agent diagnostics button and report now live in Settings -> Debug
(developer mode) instead of the Connect tab. Connect stays focused on team
connections; the diagnostics scope/state logic moves unchanged into
agent-context-diagnostics-section.tsx and is wired to DebugView by the
settings route. The connect-agent-diagnostics eval flow now drives Settings
-> Debug and proves the launcher is gone from Connect.
